refactor: split models into dedicated modules
