+++ src/models/__init__.py
@@ -1,0 +1,16 @@
+from .body import BodyMeasurement, BodyMeasurementAverages
+from .nutrition import NutritionEntry, DailyNutritionSummary, StatusResponse
+from .workout import Workout, WorkoutLog, StravaEvent, AthleteMetrics, ComplexAdvice
+
+__all__ = [
+    'BodyMeasurement',
+    'BodyMeasurementAverages',
+    'NutritionEntry',
+    'DailyNutritionSummary',
+    'StatusResponse',
+    'Workout',
+    'WorkoutLog',
+    'StravaEvent',
+    'AthleteMetrics',
+    'ComplexAdvice',
+]
